Refactor version script
This change refactors the version Python script so that it can be used
in the snap and the Python package.
