+++ tox.ini
@@ -8,4 +8,4 @@ deps =
 commands = pytest rabbitmqtesttool tests
 
 [testenv:lint]
-commands = pycodestyle --verbose tests rabbitmqtesttool test-rabbit.py setup.py
+commands = pycodestyle --verbose tests rabbitmqtesttool test-rabbit.py setup.py bin/get_version.py
